+++ xoeuf/fields/enumeration/__init__.py
@@ -12,6 +12,7 @@ from __future__ import (
     absolute_import as _py3_abs_import,
 )
 
+import logging
 from collections import namedtuple, Mapping
 from xoutil.string import cut_prefix
 from xoutil.objects import import_object
@@ -20,7 +21,10 @@ from odoo import fields, api, models
 from odoo.fields import Char
 from odoo.release import version_info as ODOO_VERSION_INFO
 
+
 from ...eight import string_types
+
+logger = logging.getLogger(__name__)
 
 
 if ODOO_VERSION_INFO < (12,):
@@ -71,6 +75,13 @@ class Enumeration(Char, _EnumeratedField):
 
     .. rubric:: Automatic selection field
 
+    .. warning:: Known issues with this features.
+
+       Up to 0.66.0 this feature is not yet stable.  The injected field is not
+       visible to ``api.depends`` and other components of Odoo framework.
+
+       However, the method `get_selection_field`:meth: seems to work well.
+
     Enumeration fields cannot be put in views because their values can be
     arbitrary Python objects which are not easily transferred to/from the web
     client.
@@ -104,12 +115,18 @@ class Enumeration(Char, _EnumeratedField):
     .. versionchanged:: 0.64.0 The `enumclass` can be a callable.  Enumeration
        is now class instead of a function.  Gain the ``Enumclass`` attribute.
 
+    .. versionchanged:: 0.66.0 The function `compute_member_string` must now
+       take three arguments: the model, the name and the value.
+
     """
+
+    type = "enumeration"
 
     _slots = {
         "enumclass": None,
         "selection_field_kwargs": None,
         "compute_member_string": None,
+        "Enumclass": None,
     }
 
     def __init__(self, enumclass, *args, **kwargs):
@@ -153,73 +170,61 @@ class Enumeration(Char, _EnumeratedField):
         return enumclass_factory(model)
 
     def get_selection_field(
-        self, field_name, name, compute_member_string=None, **kwargs
+        self, field_name, selection_field_name, compute_member_string=None, **kwargs
     ):
         """Return a computed Selection field to set/get the Enumeration.
 
         """
-        selection_field_name = name
 
         @api.multi
         @api.depends(field_name)
         def _compute_selection_field(rs):
+            enumclass = self.resolve_enumclass(rs)
             for record in rs:
                 value = getattr(record, field_name, None)
                 if value is not None:
-                    member = self.get_member_by_value(value)
+                    member = _get_member_by_value(enumclass, value)
                     setattr(record, selection_field_name, member.name)
                 else:
                     setattr(record, selection_field_name, False)
 
         @api.multi
         def _set_selection(rs):
+            enumclass = self.resolve_enumclass(rs)
             for record in rs:
                 key = getattr(record, selection_field_name, None)
                 if key:
-                    member = self.get_member_by_name(key)
+                    member = _get_member_by_name(enumclass, key)
                     setattr(record, field_name, member.value)
                 else:
                     setattr(record, field_name, False)
 
         if not compute_member_string:
-            compute_member_string = self._compute_member_string
+            compute_member_string = self._default_compute_member_string
         kwargs.setdefault("store", False)
         kwargs.setdefault("compute", _compute_selection_field)
         kwargs.setdefault("inverse", _set_selection)
         return fields.Selection(
             selection=lambda s: [
-                (name, compute_member_string(name, value))
-                for name, value in self.Enumclass.__members__.items()
+                (name, compute_member_string(s, name, value))
+                # Don't use `self.Enumclass`: when the selection field is
+                # computed the setup_full may not be called yet.
+                for name, value in self.resolve_enumclass(s).__members__.items()
             ],
             **kwargs
         )
 
     @staticmethod
-    def _compute_member_string(name, value):
+    def _default_compute_member_string(self, name, value):
+        # The 'self' is actually a parameter that (the model); it's not a typo
+        # and this function IS intentionally a staticmethod.
         return name
 
-    def get_member_by_value(self, value):
-        """Find the enumclass's member that is equal to `value`."""
-        try:
-            return next(
-                Member(k, v)
-                for k, v in self.Enumclass.__members__.items()
-                if v == value
-            )
-        except StopIteration:
-            raise ValueError(
-                "Invalid member %r of enumeration %r"
-                % (value, self.Enumclass.__members__)
-            )
+    def get_member_by_value(self, value, record=None):
+        return _get_member_by_value(self.Enumclass, value)
 
     def get_member_by_name(self, name):
-        """Find the enumclass's member by name"""
-        try:
-            return Member(name, self.Enumclass.__members__[name])
-        except (AttributeError, KeyError):
-            raise ValueError(
-                "Invalid key %r of enumeration %r" % (name, self.Enumclass.__members__)
-            )
+        return _get_member_by_name(self.Enumclass, name)
 
     def setup_full(self, model):
         # Injects a new class in the model MRO, so that we can guarantee
@@ -238,25 +243,28 @@ class Enumeration(Char, _EnumeratedField):
         # (delegated) copies for this Enumeration fields, because the
         # actual field does the right thing with DB (it does have the
         # EnumerationAdapter injected.)
-        cls = type(model)
-        if EnumerationAdapter not in cls.mro() and not self.related:
-            cls.__bases__ = (EnumerationAdapter,) + cls.__bases__
-            # I tried to use model._add_field because it changes the
-            # cls._fields that Odoo is iterating and that raises an error
-            # (RuntimeError: dictionary changed size during iteration).
-            #
-            # So let's resort to injection.
-            if self.selection_field_kwargs:
-                selection_field_name = self.selection_field_kwargs.pop("name")
+        if self._setup_done != "full":
+            cls = type(model)
+            logger.info("Setting %s to model %s(%s, %s)", self, model, cls, id(cls))
+            if EnumerationAdapter not in cls.mro() and not self.related:
+                cls.__bases__ = (EnumerationAdapter,) + cls.__bases__
+                logger.info("Setting %s to model %s(%s, %s)", self, model, cls, id(cls))
+            self.Enumclass = self.resolve_enumclass(model)
+            result = super(Enumeration, self).setup_full(model)
+            if not self.compute and not self.related and not model._abstract:
+                assert self.name in model.fields_get()
+            return result
+
+    def _add_selection_field(self, model):
+        if self.selection_field_kwargs:
+            cls = type(model)
+            selection_field_name = self.selection_field_kwargs.pop("name")
+            if selection_field_name not in cls._fields and not self.related:
                 selection_field = self.get_selection_field(
                     self.name, selection_field_name, **self.selection_field_kwargs
                 )
-                cls.__bases__ = (
-                    SelectionMixin(self.name, selection_field_name, selection_field),
-                ) + cls.__bases__
-        result = super(Enumeration, self).setup_full(model)
-        self.Enumclass = self.resolve_enumclass(model)
-        return result
+                logger.info("Adding %s to model %s", selection_field_name, model)
+                model._add_field(selection_field_name, selection_field)
 
     def convert_to_read(self, value, record, use_name_get=True):
         if value is not None and value is not False:
@@ -320,8 +328,17 @@ class EnumerationAdapter(Adapter):
             if not isinstance(query_part, string_types):
                 fieldname, operator, operands = query_part
                 field = self._fields.get(fieldname, None)
-                if isinstance(field, _EnumeratedField):
-                    if operator in ("=", "!="):
+                if isinstance(field, Enumeration):
+                    if operator in (
+                        "=",
+                        "!=",
+                        "ilike",
+                        "not ilike",
+                        "like",
+                        "not like",
+                        "=like",
+                        "=ilike",
+                    ):
                         values = _get_db_value(field, operands)
                     elif operator in ("in", "not in"):
                         values = [_get_db_value(field, o) for o in operands]
@@ -359,14 +376,6 @@ class EnumerationAdapter(Adapter):
         return super(EnumerationAdapter, self).write(values)
 
 
-def SelectionMixin(field_name, selection_field_name, selection_field):
-    return type(
-        "selection_mixin_" + field_name,
-        (Adapter,),
-        {selection_field_name: selection_field},
-    )
-
-
 def _get_db_value(field, value):
     if value is None or value is False:  # and not field.required
         return value
@@ -377,6 +386,24 @@ def _get_db_value(field, value):
         return value
 
 
+def _get_member_by_value(enumclass, value):
+    """Find the enumclass's member that is equal to `value`."""
+    try:
+        return next(
+            Member(k, v) for k, v in enumclass.__members__.items() if v == value
+        )
+    except StopIteration:
+        raise ValueError("Invalid member %r of enumeration %r" % (value, enumclass))
+
+
+def _get_member_by_name(enumclass, name):
+    """Find the enumclass's member by name"""
+    try:
+        return Member(name, enumclass.__members__[name])
+    except (AttributeError, KeyError):
+        raise ValueError("Invalid key %r of enumeration %r" % (name, enumclass))
+
+
 def constant(value):
     "Create a function (of many args) that always returns a constant `value`."
 
@@ -384,3 +411,29 @@ def constant(value):
         return value
 
     return result
+
+
+if ODOO_VERSION_INFO < (11, 0):
+
+    @api.model
+    def _setup_fields(self, partial):
+        cls = type(self)
+        for field in dict(cls._fields).values():
+            if isinstance(field, Enumeration):
+                field._add_selection_field(self)
+        _super_setup_fields(self, partial)
+
+
+else:
+
+    @api.model
+    def _setup_fields(self):
+        cls = type(self)
+        for field in dict(cls._fields).values():
+            if isinstance(field, Enumeration):
+                field._add_selection_field(self)
+        _super_setup_fields(self)
+
+
+_super_setup_fields = models.BaseModel._setup_fields
+models.BaseModel._setup_fields = _setup_fields