+++ xoeuf/tests/test_enumeration/models.py
@@ -46,23 +46,21 @@ class WORK_TYPE(object):
 
 class Mixin(models.AbstractModel):
     _name = "test_enumeration.mixin"
-    color = fields.Enumeration(
-        COLORS, default=COLORS.Red, selection_field_name="color_name"
-    )
-    color_rgb = fields.Char(compute="_compute_color_rgb")
-    color_rgb2 = fields.Char(compute="_compute_color_rgb", store=True)
+    color = fields.Enumeration(COLORS, default=COLORS.Red)
+    color_name = color.get_selection_field("color", "color_name")
+    # color_rgb = fields.Char(compute="_compute_color_rgb")
+    # color_rgb2 = fields.Char(compute="_compute_color_rgb", store=True)
 
-    @api.depends("color_name")
-    def _compute_color_rgb(self):
-        for record in self:
-            if record.color_name == "Red":
-                record.color_rgb = "#f00"
-            elif record.color_name == "Blue":
-                record.color_rgb = "#00f"
-            elif record.color_name == "Green":
-                record.color_rgb = "#0f0"
-            else:
-                record.color_rgb = "#fff"
+    # def _compute_color_rgb(self):
+    #     for record in self:
+    #         if record.color_name == "Red":
+    #             record.color_rgb = "#f00"
+    #         elif record.color_name == "Blue":
+    #             record.color_rgb = "#00f"
+    #         elif record.color_name == "Green":
+    #             record.color_rgb = "#0f0"
+    #         else:
+    #             record.color_rgb = "#fff"
 
 
 class Model(models.Model):
@@ -87,10 +85,9 @@ class Model(models.Model):
 
     dynamic_enum = fields.Enumeration(_get_enumclass)
 
-    @api.depends("color_name")
-    def _compute_color_rgb(self):
-        for record in self:
-            record.color_rgb = "#000"
+    # def _compute_color_rgb(self):
+    #     for record in self:
+    #         record.color_rgb = "#000"
 
 
 if MAJOR_ODOO_VERSION < 12: